Initial smart contract commit
Marketplace basic functions (list nft, buy nft, delete listing and changing listing price) all implemented and tested

Artist rights token implemented with a royalty system for every holder. This feature needs to be tested on the lukso blockchain with a lot of users (100,000) addresses that hold the token. More testing needs to be done with this contract

The initial sale of artist token contract is created with the creation of the artist token. Buying tokens is also implemented. Some testing is done but more will need to be added

Simple NFT contract is also created (create nft, choose its name, symbol and max supply). Mint function is also created. This contract needs features to be implemented as well as testing

Artist nft sale contract is implemented with features such as creating a sale (which also creates an NFT), mint in nft. This contract also needs work and more features to be implemented as well as testing