Added all requested features and updates
Made initial rights sale complete a bulk of its computation on creation instead of when a user calls a function on the contract (great for gas)

Added locked liquidity on exchange. As well as locking liquidity on token sales.

Made tokens refunded available to be bought again

More testing completing that suits these updates